Search Header and recent search Opt in separated into UI elements
