+++ ios/brave-ios/Sources/Brave/Frontend/Browser/Favorites/FavoritesViewController+KeyboardHelper.swift
@@ -44,27 +44,4 @@ extension FavoritesViewController: KeyboardHelperDelegate {
   ) {
     updateKeyboardInset(state)
   }
-
-  var availableSections: [FavoritesSection] {
-    var sections = [FavoritesSection]()
-    if UIPasteboard.general.hasStrings || UIPasteboard.general.hasURLs {
-      sections.append(.pasteboard)
-    }
-
-    if let favoritesObjects = favoritesFRC.fetchedObjects, !favoritesObjects.isEmpty {
-      sections.append(.favorites)
-    }
-
-    if !privateBrowsingManager.isPrivateBrowsing
-      && Preferences.Search.shouldShowRecentSearches.value,
-      RecentSearch.totalCount() > 0
-    {
-      sections.append(.recentSearches)
-    } else if !privateBrowsingManager.isPrivateBrowsing
-      && Preferences.Search.shouldShowRecentSearchesOptIn.value
-    {
-      sections.append(.recentSearches)
-    }
-    return sections
-  }
 }
